Add a route to test sending emails directly
I've been having a number of issues with the various GitHub Actions that
I've been trying to use. So, this change adds a route that has PHP code
to test (locally) if I can send with SendGrid or not, such as whether
I'm using the wrong region of an API key that lacks the required
permissions.